company db integrated

+++ app/Services/cardDetails/DestroyCardService.php
@@ -5,8 +5,10 @@ namespace App\Services\cardDetails;
 use App\Models\CardDetails;
 
 class DestroyCardService{
-    public function destroyCard($cardId){
-        $response = CardDetails::find($cardId);
-        return $response->delete();
+    public function destroyCard($data){
+        $response = CardDetails::where('id',$data[0])->where('client_id',$data[1]);
+        if($response){
+             return $response->delete();
+        }        
     }
 }
